\name{covid}
\docType{data}
\alias{covid}
\title{COVID-19 cases at the 2020 Summer Olympics and 2020 Summer Paralympics}
\usage{
data(covid)
}
\description{
  This dataset contains the number of cases from 2020 Summer Olympics and 2020 Summer Paralympics in Tokyo.
}
\format{This data frame consists of 5 variables on 70 dates:
\itemize{
  \item\code{date} Date of recording the number of cases of COVID-19.
  \item\code{athlete} Number of cases for athletes.
  \item\code{staff} Number of cases for staff.
  \item\code{volunteer} Number of cases for volunteers.
  \item\code{total} Total number of cases.
}
}
\source{Wikipedia contributors. (2022, December 5). COVID-19 cases at the 2020 Summer Olympics and 2020 Summer Paralympics. In Wikipedia, The Free Encyclopedia. Retrieved 13:01, December 20, 2022, \url{https://en.wikipedia.org/wiki/COVID-19_cases_at_the_2020_Summer_Olympics_and_2020_Summer_Paralympics#cite_note-Tokyo2020CovidList-1}
}

\keyword{datasets}
