\name{ZIDW-package}
\alias{ZIDW-package}
\alias{ZIDW}
\docType{package}
\title{
Tools for Analyzing Zero-Inflated Discrete Weibull Models
}
\description{
Functions for analyzing zero-inflated discrete Weibull (ZIDW) regression models. 
    This package includes the parameter estimations for both regression and univariate settings, 
    the distribution functions, functions to generate randomized quantile residuals and a pseudo R2,
    and plotting of rootograms.
}
\details{
\tabular{ll}{
Package: \tab ZIDW\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2025-09-15\cr
Imports: \tab DWreg, actuar, maxLik, COUNT, gtools, matrixcalc, DiscreteWeibull, dplyr, ggplot2, purrr, tibble\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Peng Yeh, Ph.D.

Derek S. Young, Ph.D.

Maintainer: Derek S. Young <derek.young@uky.edu>
}
\references{
  Kalktawi, H. S. (2017), \emph{Discrete Weibull Regression Model for Count Data}. Ph.D. Thesis, Brunel University London.

  Taconeli, C. A. and Rodrigues de Lara, I. A. (2022), Discrete Weibull Distribution: Different Estimation Methods Under Ranked Set Sampling and Simple Random Sampling. \emph{Journal of Statistical Computation and Simulation}, \bold{92}:8, 1740--1762.

  Yeh, P. and Young, D. S. (2025), Some Estimation and Inference Considerations for the Zero-Inflated Discrete Weibull Distribution. \emph{Communications in Statistics - Simulation and Computation (in press)}, 1--22.
} 

